lab01/requirements.txt
numpy
opencv-python
pillow
requests